Put each page into its own .py script
